codex - add shared docs drawer
